Move toolip logic to flatmapQueries.js
