+++ src/components/Tooltip.vue
@@ -70,7 +70,7 @@
         <external-resource-card :resources="resources"></external-resource-card>
 
         <!-- pubmed-viewer is just used for processing pubmed requests (no display) -->
-        <pubmed-viewer v-if="content.featureIds" v-show="false" :entry="content" @pubmedSearchUrl="pubmedSearchUrlUpdate"/>
+        <!-- <pubmed-viewer v-if="content.featureIds" v-show="false" :entry="content" @pubmedSearchUrl="pubmedSearchUrlUpdate"/> -->
         <el-button  v-if="pubmedSearchUrl != ''" class="button" icon="el-icon-notebook-2" @click="openUrl(pubmedSearchUrl)">
           Open publications in pubmed
         </el-button>
@@ -114,11 +114,6 @@ const inArray = function(ar1, ar2){
     return as1.indexOf(as2) !== -1
 }
 
-// remove duplicates by stringifying the objects
-const removeDuplicates = function(arrayOfAnything){
-  return [...new Set(arrayOfAnything.map(e => JSON.stringify(e)))].map(e => JSON.parse(e)) 
-}
-
 const capitalise = function(str){
   if (str)
     return str.charAt(0).toUpperCase() + str.slice(1)
@@ -158,17 +153,6 @@ export default {
       componentsWithDatasets: [],
       uberons: [{id: undefined, name: undefined}]
     };
-  },
-  inject: ['sparcAPI', 'flatmapAPI'],
-  watch: {
-    'content.featureIds': {
-      handler: function(){
-        this.pathwayQuery(this.content.featureIds)
-      }
-    }
-  },
-  mounted: function(){
-    this.getOrganCuries()
   },
   computed: {
     resources: function(){
@@ -220,204 +204,6 @@ export default {
         console.log('aborting requests')
       }
     },
-    findAllIdsFromConnectivity(connectivity){
-      let dnodes = connectivity.connectivity.flat() // get nodes from edgelist
-      let nodes = [...new Set(dnodes)] // remove duplicates
-      let found = []
-      nodes.forEach(n=>{
-        if (Array.isArray(n)){
-          found.push(n.flat())
-        } else {
-          found.push(n)
-        }
-      })
-      return [... new Set(found.flat())]
-    },
-    flattenConntectivity(connectivity){
-      let dnodes = connectivity.flat() // get nodes from edgelist
-      let nodes = [...new Set(dnodes)] // remove duplicates
-      let found = []
-      nodes.forEach(n=>{
-        if (Array.isArray(n)){
-          found.push(n.flat())
-        } else {
-          found.push(n)
-        }
-      })
-      return found.flat()
-    },
-    findComponents: function(connectivity){
-      let dnodes = connectivity.connectivity.flat() // get nodes from edgelist
-      let nodes = removeDuplicates(dnodes)
-
-      let found = []
-      let terminal = false
-      nodes.forEach(node=>{
-        terminal = false
-        // Check if the node is an destination or origin (note that they are labelled dendrite and axon as opposed to origin and destination)
-        if(inArray(connectivity.axons,node)){
-          terminal = true
-        }
-        if(inArray(connectivity.dendrites, node)){
-          terminal = true
-        }
-        if (!terminal){
-          found.push(node)
-        }
-      })
-
-      return found
-    },
-    getOrganCuries: function(){
-      fetch(`${this.sparcAPI}get-organ-curies/`)
-        .then(response=>response.json())
-        .then(data=>{
-          this.uberons = data.uberon.array
-        })
-    },
-    buildConnectivitySqlStatement: function(keastIds) {
-      let sql = 'select knowledge from knowledge where entity in ('
-      if (keastIds.length === 1) {
-        sql += `'${keastIds[0]}')`
-      } else if (keastIds.length > 1) {
-        for (let i in keastIds) {
-          sql += `'${keastIds[i]}'${i >= keastIds.length - 1 ? ')' : ','} `
-        }
-      }
-      return sql
-    },
-    buildLabelSqlStatement: function(uberons) {
-      let sql = 'select entity, label from labels where entity in ('
-      if (uberons.length === 1) {
-        sql += `'${uberons[0]}')`
-      } else if (uberons.length > 1) {
-        for (let i in uberons) {
-          sql += `'${uberons[i]}'${i >= uberons.length - 1 ? ')' : ','} `
-        }
-      }
-      return sql
-    },
-    createLabelLookup: function(uberons) {
-      return new Promise(resolve=> {
-        let uberonMap = {}
-        const data = { sql: this.buildLabelSqlStatement(uberons)}
-        fetch(`${this.flatmapAPI}knowledge/query/`, {
-            method: 'POST',
-            headers: {
-              'Content-Type': 'application/json',
-            },
-            body: JSON.stringify(data),
-          })
-          .then(response => response.json())
-          .then(payload => {
-            const entity = payload.keys.indexOf("entity");
-            const label = payload.keys.indexOf("label");
-            if (entity > -1 && label > -1) {
-              payload.values.forEach(pair => {
-                uberonMap[pair[entity]] = pair[label];
-              });
-            }
-          resolve(uberonMap)
-          })
-      })
-    },
-    pathwayQuery: function(keastIds){
-      this.controller = new AbortController();
-      const signal = this.controller.signal;
-
-      this.destinations = []
-      this.origins = []
-      this.components = []
-      this.loading = true
-      if (!keastIds || keastIds.length == 0) return
-      const data = { sql: this.buildConnectivitySqlStatement(keastIds)};
-      fetch(`${this.flatmapAPI}knowledge/query/`, {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-        body: JSON.stringify(data),
-        signal: signal
-      })
-      .then(response => response.json())
-      .then(data => {
-        if(data.values && data.values.length > 0 && JSON.parse(data.values[0][0]).connectivity && JSON.parse(data.values[0][0]).connectivity.length > 0) {
-          this.showToolipEmit(true)
-          let connectivity = JSON.parse(data.values[0][0])
-          this.processConnectivity(connectivity)
-          this.loading = false
-        } else {
-          this.showToolipEmit(false)
-        }
-        this.controller = false
-      })
-      .catch((error) => {
-        console.error('Error:', error);
-      })
-    },
-    showToolipEmit: function(bool){
-      if(!bool && this.resources.length === 0){
-        this.$emit('show-tooltip', false)
-      } else {
-        this.$emit('show-tooltip', true)
-        this.loading = false
-      }
-    },
-    createComponentsLabelList: function(components, lookUp){
-      let labelList = []
-      components.forEach(n=>{
-        labelList.push(this.createLabelFromNeuralNode(n[0]), lookUp)
-        if (n.length === 2){
-          labelList.push(this.createLabelFromNeuralNode(n[1]), lookUp)
-        }
-      })
-      return labelList
-    },
-    createLabelFromNeuralNode: function(node, lookUp){
-      let label = lookUp[node[0]]
-      if (node.length === 2 && node[1].length > 0){
-        node[1].forEach(n=>{
-          if (lookUp[n] == undefined){
-            label += `, ${n}` 
-          } else {
-            label += `, ${lookUp[n]}`
-          }
-        })
-      }
-      return label
-    },
-    processConnectivity(connectivity){
-      // Filter the origin and destinations from components
-      let components = this.findComponents(connectivity)
-
-      // Remove duplicates
-      let axons = removeDuplicates(connectivity.axons)
-      let dendrites = removeDuplicates(connectivity.dendrites)
-
-      // Create list of ids to get labels for
-      let conIds = this.findAllIdsFromConnectivity(connectivity)  
-
-      // Create readable labels from the nodes. Setting this to 'this.origins' updates the display
-      this.createLabelLookup(conIds).then(lookUp=>{
-        this.destinations = axons.map(a=>this.createLabelFromNeuralNode(a,lookUp))
-        this.origins = dendrites.map(d=>this.createLabelFromNeuralNode(d,lookUp))
-        this.components = components.map(c=>this.createLabelFromNeuralNode(c, lookUp))
-      })
-
-      this.flattenAndFindDatasets(components, axons, dendrites)
-    },
-    flattenAndFindDatasets(components, axons, dendrites){
-      
-      // process the nodes for finding datasets (Note this is not critical to the tooltip, only for the 'search on components' button)
-      let componentsFlat = this.flattenConntectivity(components)
-      let axonsFlat = this.flattenConntectivity(axons)
-      let dendritesFlat = this.flattenConntectivity(dendrites)
-
-      // Filter for the anatomy which is annotated on datasets
-      this.destinationsWithDatasets = this.uberons.filter(ub => axonsFlat.indexOf(ub.id) !== -1)
-      this.originsWithDatasets = this.uberons.filter(ub => dendritesFlat.indexOf(ub.id) !== -1)
-      this.componentsWithDatasets = this.uberons.filter(ub => componentsFlat.indexOf(ub.id) !== -1)
-    }
   }
 };
 </script>
